Uncluttered the Intermine GraphQL API file by moving model and PathQuery related content into separate files.
